feat: better handle error in end of declaration flow
